virtio-iommu: Add PROBE properties for tables
Add several PROBE properties required for attaching tables to a domain.
The MAP_UNMAP interface is simple enough that the same global properties
can be used for all endpoints managed by the device. This won't be the
case for the table interface. Hardware and emulated endpoints have
difference properties, which have to be described in PROBE requests.

* PAGE_SIZE_MASK: supported mapping granularity
* INPUT_RANGE: input address size (IOVA)
* OUTPUT_SIZE: output address size (GPA)
* PASID_SIZE: Process Address Space ID size. To know the maximum size of
  a PASID table.
* TABLE_FORMAT: PASID or page table format. These will be described in
  subsequent patches.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>